backend/requirements.txt
flask
pymongo
python-dotenv
firebase-admin
requests
google-cloud-vision
gunicorn
flask-cors
